moved external.push to gui.push
