+++ src/main/java/net/sf/jabref/gui/push/PushToApplicationAction.java
@@ -13,7 +13,7 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-package net.sf.jabref.external.push;
+package net.sf.jabref.gui.push;
 
 import java.awt.event.ActionEvent;
 import java.util.List;
