+++ src/main/java/net/sf/jabref/gui/push/PushToVim.java
@@ -13,14 +13,13 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-package net.sf.jabref.external.push;
+package net.sf.jabref.gui.push;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
 
 import javax.swing.Icon;
-import javax.swing.JCheckBox;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
@@ -31,7 +30,6 @@ import net.sf.jabref.MetaData;
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.gui.IconTheme;
 import net.sf.jabref.logic.l10n.Localization;
-import net.sf.jabref.logic.util.OS;
 import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.preferences.JabRefPreferences;
@@ -40,48 +38,44 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
 /**
- * Created by IntelliJ IDEA. User: alver Date: Jan 14, 2006 Time: 4:55:23 PM
+ * Created by IntelliJ IDEA. User: alver Date: Mar 7, 2007 Time: 6:55:56 PM To change this template use File | Settings
+ * | File Templates.
  */
-public class PushToEmacs extends AbstractPushToApplication implements PushToApplication {
+public class PushToVim extends AbstractPushToApplication implements PushToApplication {
 
-    private static final Log LOGGER = LogFactory.getLog(PushToEmacs.class);
+    private static final Log LOGGER = LogFactory.getLog(PushToVim.class);
 
-    private final JTextField additionalParams = new JTextField(30);
-    private final JCheckBox useEmacs23 = new JCheckBox();
+    private final JTextField vimServer = new JTextField(30);
 
 
     @Override
     public String getApplicationName() {
-        return "Emacs";
+        return "Vim";
     }
 
     @Override
     public Icon getIcon() {
-        return IconTheme.getImage("emacs");
+        return IconTheme.getImage("vim");
     }
 
     @Override
     public JPanel getSettingsPanel() {
-        additionalParams.setText(Globals.prefs.get(JabRefPreferences.EMACS_ADDITIONAL_PARAMETERS));
-        useEmacs23.setSelected(Globals.prefs.getBoolean(JabRefPreferences.EMACS_23));
+        vimServer.setText(Globals.prefs.get(JabRefPreferences.VIM_SERVER));
         return super.getSettingsPanel();
     }
 
     @Override
     public void storeSettings() {
         super.storeSettings();
-        Globals.prefs.put(JabRefPreferences.EMACS_ADDITIONAL_PARAMETERS, additionalParams.getText());
-        Globals.prefs.putBoolean(JabRefPreferences.EMACS_23, useEmacs23.isSelected());
+        Globals.prefs.put(JabRefPreferences.VIM_SERVER, vimServer.getText());
     }
 
     @Override
     protected void initSettingsPanel() {
         super.initSettingsPanel();
-        builder.appendRows("2dlu, p, 2dlu, p");
-        builder.add(Localization.lang("Additional parameters") + ":").xy(1, 3);
-        builder.add(additionalParams).xy(3, 3);
-        builder.add(Localization.lang("Use EMACS 23 insertion string") + ":").xy(1, 5);
-        builder.add(useEmacs23).xy(3, 5);
+        builder.appendRows("2dlu, p");
+        builder.add(Localization.lang("Vim server name") + ":").xy(1, 3);
+        builder.add(vimServer).xy(3, 3);
         settings = builder.build();
     }
 
@@ -100,33 +94,11 @@ public class PushToEmacs extends AbstractPushToApplication implements PushToAppl
             return;
         }
 
-        commandPath = Globals.prefs.get(commandPathPreferenceKey);
-        String[] addParams = Globals.prefs.get(JabRefPreferences.EMACS_ADDITIONAL_PARAMETERS).split(" ");
         try {
-            String[] com = new String[addParams.length + 2];
-            com[0] = commandPath;
-            System.arraycopy(addParams, 0, com, 1, addParams.length);
-            String prefix;
-            String suffix;
-            if (Globals.prefs.getBoolean(JabRefPreferences.EMACS_23)) {
-                prefix = "(with-current-buffer (window-buffer) (insert ";
-                suffix = "))";
-            } else {
-                prefix = "(insert ";
-                suffix = ")";
-            }
-
-            com[com.length - 1] = OS.WINDOWS ?
-            // Windows gnuclient escaping:
-            // java string: "(insert \\\"\\\\cite{Blah2001}\\\")";
-            // so cmd receives: (insert \"\\cite{Blah2001}\")
-            // so emacs receives: (insert "\cite{Blah2001}")
-            prefix.concat("\\\"\\" + getCiteCommand().replaceAll("\\\\", "\\\\\\\\") + "{" + keys + "}\\\"").concat(suffix) :
-            // Linux gnuclient escaping:
-            // java string: "(insert \"\\\\cite{Blah2001}\")"
-            // so sh receives: (insert "\\cite{Blah2001}")
-            // so emacs receives: (insert "\cite{Blah2001}")
-            prefix.concat("\"" + getCiteCommand().replaceAll("\\\\", "\\\\\\\\") + "{" + keys + "}\"").concat(suffix);
+            String[] com = new String[] {commandPath, "--servername",
+                    Globals.prefs.get(JabRefPreferences.VIM_SERVER), "--remote-send",
+                    "<C-\\><C-N>a" + getCiteCommand() +
+                    "{" + keys + "}"};
 
             final Process p = Runtime.getRuntime().exec(com);
 
@@ -143,7 +115,7 @@ public class PushToEmacs extends AbstractPushToApplication implements PushToAppl
                     }
                     // Error stream has been closed. See if there were any errors:
                     if (!sb.toString().trim().isEmpty()) {
-                        LOGGER.warn("Push to Emacs error: " + sb);
+                        LOGGER.warn("Push to Vim error: " + sb);
                         couldNotConnect = true;
                     }
                 } catch (IOException e) {
@@ -152,22 +124,25 @@ public class PushToEmacs extends AbstractPushToApplication implements PushToAppl
             });
         } catch (IOException excep) {
             couldNotCall = true;
-            LOGGER.warn("Problem pushing to Emacs.", excep);
+            LOGGER.warn("Problem pushing to Vim.", excep);
         }
+
     }
 
     @Override
     public void operationCompleted(BasePanel panel) {
         if (couldNotConnect) {
-            JOptionPane.showMessageDialog(panel.frame(), "<HTML>" +
-                    Localization.lang("Could not connect to a running gnuserv process. Make sure that "
-                            + "Emacs or XEmacs is running,<BR>and that the server has been started "
-                            + "(by running the command 'server-start'/'gnuserv-start').") + "</HTML>",
+            JOptionPane.showMessageDialog(
+                    panel.frame(),
+                    "<HTML>" +
+                            Localization.lang("Could not connect to Vim server. Make sure that "
+                                    + "Vim is running<BR>with correct server name.")
+                    + "</HTML>",
                     Localization.lang("Error"), JOptionPane.ERROR_MESSAGE);
         } else if (couldNotCall) {
-            JOptionPane.showMessageDialog(panel.frame(),
-                    Localization.lang("Could not run the gnuclient/emacsclient program. Make sure you have "
-                            + "the emacsclient/gnuclient program installed and available in the PATH."),
+            JOptionPane.showMessageDialog(
+                    panel.frame(),
+                    Localization.lang("Could not run the 'vim' program."),
                     Localization.lang("Error"), JOptionPane.ERROR_MESSAGE);
         } else {
             super.operationCompleted(panel);
@@ -176,12 +151,7 @@ public class PushToEmacs extends AbstractPushToApplication implements PushToAppl
 
     @Override
     protected void initParameters() {
-        commandPathPreferenceKey = JabRefPreferences.EMACS_PATH;
-    }
-
-    @Override
-    protected String getCommandName() {
-        return "gnuclient " + Localization.lang("or") + " emacsclient";
+        commandPathPreferenceKey = JabRefPreferences.VIM;
     }
 
 }
